refactor: LocalFileStorage 파일 위치 및 의미있는 이름 - LocalProductVideoStorage로 변경
